[tests][python] move tests that use `train()` function defined in `engine.py` from `test_basic.py` to `test_engine.py` (#5034)
* Update test_basic.py

* Update test_engine.py

* Update test_engine.py
